# Add created_at indexes for faster get_feed.