@stereosteve 2022-03-07
After looking more at https://github.com/AudiusProject/audius-protocol/blob/master/discovery-provider/src/tasks/index_aggregate_user.py#L20

I agree with Ray's suggestion from Slack about not omitting `is_delete` rows from index.

We could maybe consider a compound index like `(owner_id, is_delete)` if we still want to have index only scans for some queries...